// modals/ModalWrapper/enums.ts
enum USER_MODAL_TYPES {
  CHECK_YOUR_EMAIL = 'check_your_email',
  CREATE_AN_ACCOUNT = 'create_an_account',
  FORGOT_PASSWORD = 'forgot_password',
  INVITE_MEMBERS = 'invite_members',
  JOIN_A_POOL = 'join_a_pool',
  RESET_PASSWORD = 'reset_password',
  SIGN_IN = 'sign_in',
  DASHBOARD_MEMBERSHIP_DETAILS = 'dashboard_membership_details',
}

enum POOL_MODAL_TYPES {
  GOLF_GROUPS_TOURNAMENTS = 'golf_groups_tournaments',
  MAKE_A_PICK_SUCCESS = 'make_a_pick_success',
}

export { POOL_MODAL_TYPES, USER_MODAL_TYPES }